move all item classes to same file
