+++ src/main/kotlin/payload/response/ItemRetrieved.kt
@@ -1,0 +1,9 @@
+package payload.response
+
+import kotlinx.serialization.Serializable
+import payload.response.common.Item
+
+@Serializable
+data class ItemRetrieved private constructor(
+	val item: Item
+)
