+++ src/main/kotlin/payload/response/Items.kt
@@ -1,17 +1,9 @@
 package payload.response
 
-import enums.*
 import kotlinx.serialization.Serializable
+import payload.response.common.ItemShort
 
 @Serializable
 data class Items private constructor(
 	val items: List<ItemShort>
-) {
-	@Serializable
-	data class ItemShort private constructor(
-		val url_name: ItemUrlName,
-		val id: IdItem,
-		val item_name: String,
-		val thumb: ResourceLocation
-	)
-}
+)
